Add mineable.
TODO: Need to add game* to the onX functions, or maybe not, because this means that they can do some werid stuff in those functions. I will have to find a way to maybe either fix that or add I queue function to anything that can modify the game. For instance removing and adding entities.